update to VS2026 fix a keng tool date convert bug(in GetTimeDesc())

+++ thprac/src/thprac/thprac_igi_key_render.h
@@ -20,5 +20,5 @@ void RecordKey(int ver, uint32_t cur_key);
 void KeysHUD(int ver, ImVec2 render_pos_arrow, ImVec2 render_pos_key, const KeyRectStyle& style, bool align_right_arrow = true, bool align_right_key = false); // assume game window's 1280*960
 void SaveKeyRecorded();
 void ClearKeyRecord();
-std::vector<uint8_t>& GetKeyAPS();
+std::vector<int>& GetKeyAPS();
 }
